Fixed merge conflict
Fixed a merge conflict made by more recent changes to fix a bug (https://github.com/house-of-abbey/GarminHomeAssistant/pull/164)

+++ source/Settings.mc
@@ -70,25 +70,29 @@ class Settings {
         // Manage this inside the application or widget only (not a glance or background service process)
         if (mIsApp) {
             if (mHasService) {
-                mWebhookManager = new WebhookManager();
-                if (getWebhookId().equals("")) {
-                    // System.println("Settings update(): Doing full webhook & sensor creation.");
-                    mWebhookManager.requestWebhookId();
-                } else {
-                    // System.println("Settings update(): Doing just sensor creation.");
-                    // We already have a Webhook ID, so just enable or disable the sensor in Home Assistant.
-                    mWebhookManager.registerWebhookSensors();
-                }
-                if (mIsSensorsLevelEnabled) {
-                    // Create the timed activity
-                    if ((Background.getTemporalEventRegisteredTime() == null) or
-                        (Background.getTemporalEventRegisteredTime() != (mBatteryRefreshRate * 60))) {
-                        Background.registerForTemporalEvent(new Time.Duration(mBatteryRefreshRate * 60)); // Convert to seconds
-                        Background.registerForActivityCompletedEvent();
+                if (System.getDeviceSettings().phoneConnected) {
+                    mWebhookManager = new WebhookManager();
+                    if (getWebhookId().equals("")) {
+                        // System.println("Settings update(): Doing full webhook & sensor creation.");
+                        mWebhookManager.requestWebhookId();
+                    } else {
+                        // System.println("Settings update(): Doing just sensor creation.");
+                        // We already have a Webhook ID, so just enable or disable the sensor in Home Assistant.
+                        mWebhookManager.registerWebhookSensors();
                     }
-                } else if (Background.getTemporalEventRegisteredTime() != null) {
-                    Background.deleteTemporalEvent();
-                    Background.deleteActivityCompletedEvent();
+                    if (mIsSensorsLevelEnabled) {
+                        // Create the timed activity
+                        if ((Background.getTemporalEventRegisteredTime() == null) or
+                            (Background.getTemporalEventRegisteredTime() != (mBatteryRefreshRate * 60))) {
+                            Background.registerForTemporalEvent(new Time.Duration(mBatteryRefreshRate * 60)); // Convert to seconds
+                            Background.registerForActivityCompletedEvent();
+                        }
+                    } else if (Background.getTemporalEventRegisteredTime() != null) {
+                        Background.deleteTemporalEvent();
+                        Background.deleteActivityCompletedEvent();
+                    }
+                } else {
+                    ErrorView.show(WatchUi.loadResource($.Rez.Strings.NoPhone) as Lang.String);
                 }
             } else {
                 // Explicitly disable the background event which persists when the application closes.
